% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{assay_to_plate_df}
\alias{assay_to_plate_df}
\title{Assay to plate dataframe}
\usage{
assay_to_plate_df(assay_df, plate_n, feature)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{plate_n}{integer, plate number to generate dataframe for}

\item{feature}{feature to generate the plate data format for eg \code{types}}
}
\value{
dataframe, plate dataframe in 96-well plate format
}
\description{
Convert full assay dataframe to 96-well plate format for a specified plate number
}
\examples{
assay_to_plate_df()
}
\keyword{assay}
