% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_2_qc.R
\name{assay_to_prism_df}
\alias{assay_to_prism_df}
\title{Assay to PRISM dataframe}
\usage{
assay_to_prism_df(assay_df, feature = "neutralisation")
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{feature}{feature to generate the plate data format for (default = \code{neutralisation})}
}
\value{
dataframe, PRISM dataframe in 96-well plate format for all plates
}
\description{
Convert full assay dataframe to an exportable format for PRISM for the specified feature
}
\keyword{PRISM}
\keyword{assay}
