% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplate-package.R
\docType{package}
\name{autoplate-package}
\alias{autoplate}
\alias{autoplate-package}
\title{autoplate: AutoPlate}
\description{
AutoPlate is an R Shiny web application (and UI) that helps you automate the analysis of biological assays conducted on 96-well plates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PhilPalmer/AutoPlate}
  \item Report bugs at \url{https://github.com/PhilPalmer/AutoPlate/issues}
}

}
\author{
\strong{Maintainer}: Phil Palmer \email{pp502@cam.ac.uk}

}
\keyword{internal}
