% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{calc_neut}
\alias{calc_neut}
\title{Calculate neutralisation}
\usage{
calc_neut(assay_df)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}
}
\value{
dataframe, containing the initialised neutralisation column
}
\description{
Calculate neutralisation using the RLU values and normalisation with the virus and cell only controls
}
\keyword{assay}
