% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{create_feature_dropdown}
\alias{create_feature_dropdown}
\title{Create feature dropdown}
\usage{
create_feature_dropdown(new_feature, input, values)
}
\arguments{
\item{new_feature}{character, name of feature to update eg "bleed"}

\item{input}{object, R shiny server input parameter/object containing \code{plate_data} dataframe and \code{new_feature} character}

\item{values}{object, containing \code{assay_df} biological assay data from plate reader}
}
\value{
dropdown of existing columns that can be used to define the \code{new_feature}
}
\description{
Create feature dropdown that can be used to select an existing column from a dataframe
}
\keyword{assay}
