% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{create_feature_table}
\alias{create_feature_table}
\title{Create feature table}
\usage{
create_feature_table(new_feature, input, values)
}
\arguments{
\item{new_feature}{character, name of feature to update eg "bleed"}

\item{input}{object, R shiny server input parameter/object containing \code{new_feature} character}

\item{values}{object, containing \code{assay_df} biological assay data from plate reader}
}
\value{
rhandsontable interactive table of levles for selected feature/table in dataframe
}
\description{
Create feature table that can be used to display existing levels for a selected column in a dataframe
}
\keyword{assay}
