% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_results.R
\name{cv_boxplot_code}
\alias{cv_boxplot_code}
\title{Cell virus boxplot code}
\usage{
cv_boxplot_code(code)
}
\arguments{
\item{code}{character, that's required, either "all" for setup and plot code or "plot" for just plot code}

\item{drm_string}{character, specifying the code the dose response model}
}
\value{
character, containing the code required for the cell virus boxplot results code
}
\description{
Function to return the cell virus boxplot results plot code
}
\examples{
cv_boxplot_code()
}
\keyword{code}
\keyword{plot}
