% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{data_exploration_code}
\alias{data_exploration_code}
\title{Data exploration code}
\usage{
data_exploration_code(code)
}
\arguments{
\item{code}{character, code that's required, either "all" for setup and plot code or "plot" for just plot code}
}
\value{
character, containing the code required for the data exploration plot results code
}
\description{
Function to return the setup code for the data exploration results plot
}
\keyword{code}
\keyword{plot}
