% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dilutions}
\alias{dilutions}
\title{Default dilutions data}
\format{
A data frame with 8 rows and 2 variables
}
\usage{
data(dilutions)
}
\description{
A dataset containing the default dilutions serum and control dilutions
}
\keyword{datasets}
