% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dilutions_ella}
\alias{dilutions_ella}
\title{Default dilutions ELLA data}
\format{
A data frame with 10 rows and 1 variable
}
\usage{
data(dilutions_ella)
}
\description{
A dataset containing the default ELLA serum dilutions
}
\keyword{datasets}
