% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dilutions_pmn}
\alias{dilutions_pmn}
\title{Default dilutions pMN data}
\format{
A data frame with 8 rows and 2 variables
}
\usage{
data(dilutions_pmn)
}
\description{
A dataset containing the default pMN serum and control dilutions
}
\keyword{datasets}
