% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_results.R
\name{drc_code}
\alias{drc_code}
\title{DRC code}
\usage{
drc_code(code, drm_string)
}
\arguments{
\item{code}{character, code that's required, either "all" for setup and plot code or "plot" for just plot code}

\item{drm_string}{character, specifying the code the dose response model}
}
\value{
character, containing the code required for the DRC plot results code
}
\description{
Function to return the DRC plot results plot code
}
\examples{
drc_code()
}
\keyword{code}
\keyword{plot}
