% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{ella_to_assay_df}
\alias{ella_to_assay_df}
\title{ELLA to assay dataframe}
\usage{
ella_to_assay_df(filename, filepath)
}
\arguments{
\item{filename}{character, file name to be added to the dataframe}

\item{filepath}{character, path to the CSV file}
}
\value{
dataframe, generated from the input Excel file containing the filenames
}
\description{
Read ELLA data from an Excel file and convert to assay dataframe format appending the filename as an additional column
}
\keyword{Excel}
\keyword{files}
\keyword{input}
\keyword{read}
