% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_column_descriptions}
\alias{example_data_column_descriptions}
\title{Example data column descriptions}
\format{
A data frame containing 13 rows and 3 variables
}
\usage{
data(example_data_column_descriptions)
}
\description{
A dataset containing an example data column descriptions
}
\keyword{datasets}
