% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_pmn_plate_n1}
\alias{example_data_pmn_plate_n1}
\title{pMN Plate 1 Example Data}
\format{
A data frame with 1152 rows and 14 variables
}
\usage{
data(example_data_pmn_plate_n1)
}
\description{
A dataset containing an example pMN input plate data
}
\keyword{datasets}
