% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_pmn_platelist_H1N1}
\alias{example_data_pmn_platelist_H1N1}
\title{pMN Platelist H1N1 Example Data}
\format{
A data frame with 1152 rows and 14 variables, see more info on the variables \href{https://philpalmer.github.io/AutoPlate/articles/web_app.html#columns-1}{here}
}
\usage{
data(example_data_pmn_platelist_H1N1)
}
\description{
A dataset containing example pMN platelist data for influenza H1N1
}
\keyword{datasets}
