% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_2_qc.R
\name{exclude_wells}
\alias{exclude_wells}
\title{Exclude wells}
\usage{
exclude_wells(assay_df, exclusion_string)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}

\item{exclusion_string}{character, comma-seperated string containing exclusion criteria}
}
\value{
dataframe, main assay with updated values in the exclude column based on the exclusion criteria
}
\description{
Exclude values in dataframe based on exclusion criteria
}
\keyword{exclude}
\keyword{wells}
