% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{ic50_boxplot_code}
\alias{ic50_boxplot_code}
\title{IC50 boxplot code}
\usage{
ic50_boxplot_code(code, drm_string, ic50_is_boxplot, virus)
}
\arguments{
\item{code}{character, code that's required, either "all" for setup and plot code or "plot" for just plot code}

\item{drm_string}{character, specifying the code the dose response model}

\item{ic50_is_boxplot}{bool, specify plot type, true for boxplot, false for scatter plot}

\item{virus}{character, specifying the virus to plot}
}
\value{
character, containing the code required for the IC50 plot results code
}
\description{
Function to return the IC50 plot results plot code
}
\keyword{code}
\keyword{plot}
