% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{init_assay_df}
\alias{init_assay_df}
\title{Initialise assay dataframe}
\usage{
init_assay_df(luminescence_files, assay_type = "pMN")
}
\arguments{
\item{luminescence_files}{dataframe, containing columns "name" and "datapath" for input file names and path to the file respectively}

\item{assay_type}{character, type of assay eg "pMN" or "ELLA" (default = "pMN")}
}
\value{
dataframe, generated from the files containing the filenames
}
\description{
Read input files spectified from a dataframe and convert to assay dataframe format appending the filename as an additional column
}
\keyword{assay}
\keyword{dataframe}
\keyword{files}
\keyword{input}
\keyword{read}
