% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_2_qc.R
\name{init_av_lum_df}
\alias{init_av_lum_df}
\title{Initialise average luminescence dataframe}
\usage{
init_av_lum_df(assay_df)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}
}
\value{
dataframe, containing average luminescence data for each plate
}
\description{
Initialise average luminescence dataframe
}
\examples{
init_av_lum_df()
}
\keyword{average}
\keyword{dataframe}
\keyword{luminescence}
