% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input.R
\name{init_cols}
\alias{init_cols}
\title{Init cols}
\usage{
init_cols(assay_df)
}
\arguments{
\item{assay_df}{Dataframe containing biological assay data from plate reader}
}
\value{
A dataframe containing the initialised columns
}
\description{
Initialise columns for assay dataframe
}
\examples{
init_cols()
}
\keyword{assay}
