% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input.R
\name{init_neut}
\alias{init_neut}
\title{Init neut}
\usage{
init_neut(assay_df)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}
}
\value{
dataframe, containing the initialised neutralisation column
}
\description{
Initialise/calculate neutralisation using the RLU values and normalisation with the virus and cell only controls
}
\examples{
init_neut()
}
\keyword{assay}
