% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{init_samples}
\alias{init_samples}
\title{Init samples}
\usage{
init_samples(assay_df, assay_type)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{assay_type}{character, type of assay eg "pMN" or "ELLA"}
}
\value{
dataframe, containing the initialised samples in sample_id column
}
\description{
Initialise all sample_ids (mouse numbers) for assay dataframe based on assay type
}
\keyword{assay}
