% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input.R
\name{init_subject}
\alias{init_subject}
\title{Init subject}
\usage{
init_subject(assay_df, wcols, subject)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{wcols}{vector, list of column numbers to update for each plate}

\item{subject}{character, name to set subject to for specified columns}
}
\value{
dataframe, containing the initialised subject column
}
\description{
Initialise subject (mouse number) for assay dataframe based on input parameters
}
\examples{
init_subject()
}
\keyword{assay}
