% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{init_title}
\alias{init_title}
\title{Initialise title}
\usage{
init_title(assay_df, full_title = FALSE)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}

\item{full_title}{bool, to specify if the full length title should be used}
}
\value{
character, title for plot
}
\description{
Function to return the title for a given assay dataframe
}
\keyword{plot}
\keyword{title}
