% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{init_types}
\alias{init_types}
\title{Init types}
\usage{
init_types(assay_df, assay_type = "pMN")
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{assay_type}{character, type of assay eg "pMN" or "ELLA" (default = "pMN")}
}
\value{
dataframe, containing the initialised type column
}
\description{
Initialise types (c,m,v,x) for assay dataframe based on template 96-well plate format
}
\keyword{assay}
