% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_2_qc.R
\name{init_types_boxplot}
\alias{init_types_boxplot}
\title{Initialise types boxplot}
\usage{
init_types_boxplot(assay_df)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}
}
\value{
boxplot containing types data for each of the different plates
}
\description{
Generate types boxplot
}
\keyword{boxplot}
\keyword{types}
