% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{plot_cv_boxplot}
\alias{plot_cv_boxplot}
\title{Plot cell-virus boxplot}
\usage{
plot_cv_boxplot(assay_df, text_size = 12)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}

\item{text_size}{integer, base text size to be used for the plot (default = 12)}
}
\value{
plot, cell-virus boxplot ggplot2 object
}
\description{
Function to generate a cell-virus boxplot plot from assay data frame
}
\keyword{boxplot}
\keyword{cell}
\keyword{virus}
