% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{plot_data_exploration}
\alias{plot_data_exploration}
\title{Plot data exploration}
\usage{
plot_data_exploration(assay_df)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}
}
\value{
plot, data exploration plot ggplot2 object
}
\description{
Function to generate a data exploration plot from assay data frame
}
\keyword{data}
\keyword{exploration}
\keyword{plot}
