% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{plot_drc}
\alias{plot_drc}
\title{Plot DRC}
\usage{
plot_drc(assay_df, drm)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}

\item{drm}{DRM object, the dose response model}
}
\value{
plot, dose-response curve plot ggplot2 object
}
\description{
Function to generate a dose-response curve (DRC) plot from assay data frame
}
\keyword{drc}
\keyword{plot}
