% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_2_qc.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap}
\usage{
plot_heatmap(plate_number, assay_df, feature, title)
}
\arguments{
\item{plate_number}{integer, to specify plate number you wish to plot from assay dataframe}

\item{assay_df}{dataframe, main assay dataframe}

\item{feature}{character, name of feature to plot, must be a column in assay_df}

\item{title}{character, title to be used for the plot title along with the plate number}
}
\value{
heatmap plot
}
\description{
Generate a heatmap plot in 96-well plate format for a feature from the assay dataframe
}
\examples{
plot_heatmap()
}
\keyword{heatmap}
\keyword{plot}
