% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{plot_ic50_boxplot}
\alias{plot_ic50_boxplot}
\title{Plot IC50 boxplot}
\usage{
plot_ic50_boxplot(assay_df, drm, plot_type = "boxplot")
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}

\item{drm}{DRM object, the dose response model}

\item{plot_type}{character, either "boxplot" (default) or "jitter"}
}
\value{
plot, dose-response curve plot ggplot2 object
}
\description{
Function to gneerate IC50 plot from input main assay dataframe and dose-response model (DRM)
}
\keyword{code}
\keyword{plot}
