% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmn_platelist_H1N1_example_data}
\alias{pmn_platelist_H1N1_example_data}
\title{pMN Platelist H1N1 Example Data}
\format{
A data frame with 1152 rows and 20 variables
}
\usage{
data(pmn_platelist_H1N1_example_data)
}
\description{
A dataset containing example pMN platelist data for influenza H1N1
}
\keyword{datasets}
