% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{prism_add_tags}
\alias{prism_add_tags}
\title{Prism add tags}
\usage{
prism_add_tags(code, language = "r")
}
\arguments{
\item{code}{character, HTML code to add the code tags to}

\item{language}{character, programming language of the HTML code block, default is R}
}
\value{
A HTML block with relevant langugage code tags
}
\description{
Add code HTML tags to a HTML block input
}
\examples{
prism_add_tags()
}
\keyword{HTML}
\keyword{code}
