% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{prism_code_block}
\alias{prism_code_block}
\title{Prism code block}
\usage{
prism_code_block(code, language = "r")
}
\arguments{
\item{code}{character, HTML code to highlight with Prism}

\item{language}{character, programming language of the HTML code block, default is R}
}
\value{
A HTML block with syntax highlighting
}
\description{
Add syntax highlighting to a HTML code block with Prism
}
\keyword{HTML}
\keyword{Prism}
\keyword{code}
