% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{read_plus}
\alias{read_plus}
\title{Read plus}
\usage{
read_plus(filename, filepath)
}
\arguments{
\item{filename}{character, file name to be added to the dataframe}

\item{filepath}{character, path to the CSV file}
}
\value{
dataframe, generated from the input CSV containing the filenames
}
\description{
Read CSV file appending the filename as an additional column in the dataframe
}
\examples{
read_plus()
}
\keyword{CSV}
\keyword{files}
\keyword{input}
\keyword{read}
