% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{replace_names}
\alias{replace_names}
\title{Replace names}
\usage{
replace_names(df, oldnames, newnames)
}
\arguments{
\item{df}{dataframe, a dataframe containing all of the old column names}

\item{oldnames}{vector, list of column names to be replaced. Must have the same order and length as new names}

\item{newnames}{vector, list of column names to replaced the old ones. Must have the same order and length as old names}
}
\value{
dataframe, dataframe with updated column names
}
\description{
Out with the old in with the new
}
\keyword{column}
\keyword{names}
\keyword{replace}
