% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_results.R
\name{setup_code}
\alias{setup_code}
\title{Setup code}
\usage{
setup_code(drc_plot)
}
\arguments{
\item{drc_plot}{bool, value to specify if the plot requires the drc library}
}
\value{
character, containing the code required to setup a results plot
}
\description{
Function to return the setup code for a results plot
}
\examples{
setup_code()
}
\keyword{code}
\keyword{setup}
