% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{update_cols_order}
\alias{update_cols_order}
\title{Update treatments colours and order}
\usage{
update_cols_order(assay_df)
}
\arguments{
\item{assay_df}{dataframe, main assay dataframe}
}
\value{
list, containing a modified assay dataframe, a list of treatment colours and a list of the treatments
}
\description{
Function to return to update the colours and ordering of treatments before plotting
}
\keyword{code}
\keyword{colours}
