% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{update_dilutions}
\alias{update_dilutions}
\title{Update dilutions}
\usage{
update_dilutions(assay_df, dilutions)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{dilutions}{dataframe, containing two columns for serum and control dilutions}
}
\value{
dataframe, containing the updated dilutions
}
\description{
Update dilutions in main assay dataframe based on the input dilutions dataframe
}
\examples{
update_dilutions()
}
\keyword{assay}
