% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{update_feature}
\alias{update_feature}
\title{Update feature}
\usage{
update_feature(new_feature, input, values)
}
\arguments{
\item{new_feature}{character, name of feature to update eg "bleed"}

\item{input}{object, R shiny server input parameter/object containing new_feature\verb{character and}new_feature_table` dataframe}

\item{values}{object, containing \code{assay_df} biological assay data from plate reader}
}
\value{
null, automatically updates the main assay dataframe within the reactive values
}
\description{
For a given feature update the values in the main assay dataframe reactive values based on the input from the feature table
}
\keyword{assay}
