% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{update_feature_plate}
\alias{update_feature_plate}
\title{Update types}
\usage{
update_feature_plate(assay_df, feature, plate_n, changes)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{feature}{character, containing column to update in assay dataframe}

\item{plate_n}{integer, plate number to update}

\item{changes}{vector, containg sub vectors each corresponding to a change, which contain 0-index based: c(row, col, old_val, new_val)}
}
\value{
dataframe, containing the updated feature
}
\description{
Update types in main assay dataframe for a single plate based on the input plate dataframe
}
\keyword{assay}
