% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{update_sample_ids}
\alias{update_sample_ids}
\title{Update sample_ids}
\usage{
update_sample_ids(assay_df, updated_plate_df, plate_n)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{updated_plate_df}{dataframe, containing 96-well plate data}

\item{plate_n}{integer, plate number to update}
}
\value{
dataframe, containing the updated sample_ids
}
\description{
Update sample_ids in main assay dataframe for a single plate based on the input plate dataframe
}
\examples{
update_sample_ids()
}
\keyword{assay}
