% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input.R
\name{update_subjects}
\alias{update_subjects}
\title{Update subjects}
\usage{
update_subjects(assay_df, updated_plate_df, plate_n)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{updated_plate_df}{dataframe, containing 96-well plate data}

\item{plate_n}{integer, plate number to update}
}
\value{
dataframe, containing the updated subjects
}
\description{
Update subjects in main assay dataframe for a single plate based on the input plate dataframe
}
\examples{
update_subjects()
}
\keyword{assay}
