% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input.R
\name{update_types}
\alias{update_types}
\title{Update types}
\usage{
update_types(assay_df, updated_plate_df, plate_n)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{updated_plate_df}{dataframe, containing 96-well plate data}

\item{plate_n}{integer, plate number to update}
}
\value{
dataframe, containing the updated types
}
\description{
Update types in main assay dataframe for a single plate based on the input plate dataframe
}
\examples{
update_types()
}
\keyword{assay}
