% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_pmn_plate_n1}
\alias{example_data_pmn_plate_n1}
\title{pMN Plate 1 Example Data}
\format{
A data frame containing 96 rows and 2 variables, see more info on the variables \href{https://philpalmer.github.io/AutoPlate/articles/web_app.html#columns-1}{here}
}
\usage{
data(example_data_pmn_plate_n1)
}
\description{
A dataset containing an example pMN input plate data
}
\keyword{datasets}
