% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_input.R
\name{init_cols}
\alias{init_cols}
\title{Init cols}
\usage{
init_cols(assay_df)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}
}
\value{
dataframe, containing the initialised columns
}
\description{
Initialise columns for assay dataframe
}
\examples{
init_cols()
}
\keyword{assay}
