% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input.R
\name{init_sample}
\alias{init_sample}
\title{Init sample ID}
\usage{
init_sample(assay_df, wcols, sample_id)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{wcols}{vector, list of column numbers to update for each plate}

\item{sample_id}{character, name to set sample_id to for specified columns}
}
\value{
dataframe, containing the initialised sample_id column
}
\description{
Initialise sample_id (mouse number) for assay dataframe based on input parameters
}
\examples{
sample()
}
\keyword{assay}
