% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_results.R
\name{setup_code}
\alias{setup_code}
\title{Setup code}
\usage{
setup_code(is_drc_plot, is_vc_plot)
}
\arguments{
\item{is_drc_plot}{bool, value to specify if the plot requires the drc library}

\item{is_vc_plot}{bool, value to specify if the plot requires the metafolio library}
}
\value{
character, containing the code required to setup a results plot
}
\description{
Function to return the setup code for a results plot
}
\examples{
setup_code()
}
\keyword{code}
\keyword{setup}
