% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input.R
\name{update_feature_plate}
\alias{update_feature_plate}
\title{Update types}
\usage{
update_feature_plate(assay_df, feature, plate_n, changes)
}
\arguments{
\item{assay_df}{dataframe, containing biological assay data from plate reader}

\item{plate_n}{integer, plate number to update}

\item{changes}{vector, containg sub vectors each corresponding to a change, which contain 0-index based: c(row, col, old_val, new_val)}

\item{updated_plate_df}{dataframe, containing 96-well plate data}
}
\value{
dataframe, containing the updated feature
}
\description{
Update types in main assay dataframe for a single plate based on the input plate dataframe
}
\examples{
update_feature_plate()
}
\keyword{assay}
